requirements_production.txt
# requirements_production.txt - Dependencias para Producción
# Versión de producción de E-Garage

# Core Django
Django==5.1.6
django-allauth==65.3.0
django-autocomplete-light==3.11.0

# Base de datos
psycopg2-binary==2.9.9  # PostgreSQL adapter
redis==5.2.1            # Cache y sesiones

# Servidor web
gunicorn==23.0.0        # WSGI server
whitenoise==6.8.2       # Static files serving

# Monitoreo y logging
sentry-sdk==2.19.2      # Error tracking
django-extensions==3.2.3 # Management commands

# Seguridad
django-cors-headers==4.6.0  # CORS handling
django-csp==3.8.0           # Content Security Policy

# Utilidades
python-decouple==3.8     # Environment variables
django-crispy-forms==2.3 # Form rendering
crispy-bootstrap4==2024.10

# Email
django-ses==4.1.1       # AWS SES backend (opcional)

# Backup y compresión
django-dbbackup==4.2.0  # Database backup
boto3==1.35.84          # AWS S3 para backups remotos (opcional)

# Análisis y reportes
django-analytical==3.1.0 # Google Analytics integration
reportlab==4.2.5         # PDF generation
openpyxl==3.1.5          # Excel export

# Performance
django-cache-panel==0.1   # Cache debugging
django-debug-toolbar==4.4.6  # Development debugging (solo dev)

# API (si se necesita)
djangorestframework==3.15.2
django-filter==24.3

# Task queue (opcional para trabajos pesados)
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Archivos y media
Pillow==11.0.0          # Image processing
django-storages==1.14.4 # S3/cloud storage (opcional)

# Testing y QA (desarrollo/CI)
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
selenium==4.27.1

# Linting y formateo (desarrollo)
black==24.10.0
flake8==7.1.1
isort==5.13.2
pre-commit==4.0.1

# Deployment
fabric==3.2.2          # Deployment automation
invoke==2.2.0          # Task runner

# Health checks
django-health-check==3.18.3

# Scheduled tasks
django-crontab==0.7.1   # Cron jobs management

# Logging
django-structlog==8.1.0  # Structured logging
colorlog==6.8.2          # Colored logs
